requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Werkzeug==2.3.7
pymongo==4.4.1
python-dotenv==1.0.0